backend/pyproject.toml
[tool.poetry]
name = "odapclean-backend"
version = "0.1.0"
description = "Backend for OdapClean"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-dotenv = "^1.0.1"
supabase = "^2.0.0"
python-multipart = "^0.0.9"
gunicorn = "^22.0.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
requests = "^2.32.5"
opencv-python-headless = "<4.10"
numpy = "<2"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"